safari-600.7-branch/LayoutTests/fast/dom/shadow/shadow-hierarchy-exception-expected.txt

Appending an ancestor to a shadow tree should throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowDiv.appendChild(container) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS successfullyParsed is true

TEST COMPLETE
